backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
google-genai
